Add `error_codes` to the ABI (#31)
# Description

This PR adds the `errorCodes` section to the ABI, as specified in the
[ABI Errors
RFC](https://github.com/FuelLabs/sway-rfcs/blob/master/rfcs/0014-abi-errors.md#revert-codes):

```
"errorCodes": {
  "18446744069414584320": {
    "pos": {
      "pkg": "my_lib",
      "file": "lib.rs",
      "line": 42,
      "column": 13
    },
    "logId": null,
    "msg": "Error message."
  },
  "18446744069414584321": {
    "pos": {
      "pkg": "my_contract",
      "file": "main.rs",
      "line": 21,
      "column": 34
    },
    "logId": "4933727799282657266",
    "msg": null
  }
}
```

The proposed structure differs slightly from the original proposal, as
it adds the `pkg` field. The intention of this field is to clearly
distinguish between the package in which the error occurs and the file
path within the package.

Separating the package from the file path:
- gives clear distinction in which package the error occurs.
- supports the cases where the package name as defined in the
`Forc.toml` does not need to be the same as the directory name. Having
the package name as a part of the file path would in such cases be
misleading and confusing.
- opens the possibility of having also the package version encoded, once
we have it available in the compiler. E.g. "my_lib@1.2.0".