Fix bug in `RevertInfo::new` (#40)
This PR fixes a bug in `RevertInfo::new` when a revert code was not a
know error signal or caused by panicking.

The bug was caused by internally returning zero for an error code to
look up, when the revert code was not a panic. But in the ABI v1.2 zero
is a valid and existing error code. The fix uses `None` instead of zero
to properly propagate the information that the error code does not
exists or in other words, that the revert code is not caused by
panicking.

The PR increases the patch version of the package to `0.15.1`.